expand shortest_queue routing heuristic to transport resources
If next resources are not production-resources, we have either sinks or (mainly) transportresources. For these, the queue can be can be obtained by checking their controllers requests. For sinks there is no need to sort as there will likely be just one, otherwise the controller requests are of length 0 because there are no open requests for sources as these take no time to process their inputs. Hence, the shuffeling results in a random sink allocation.